wallet_core/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.3.0
bitarray==3.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
ckzg==2.1.1
colorama==0.4.6
cryptography==45.0.5
cytoolz==1.0.1
defusedxml==0.7.1
Django==5.2.4
django-extension==0.1.0
django-extensions==4.1
django-ratelimit==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
dnspython==2.7.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
frozenlist==1.7.0
hexbytes==1.3.1
idna==3.10
iniconfig==2.1.0
multidict==6.6.3
oauthlib==3.3.1
packaging==25.0
parsimonious==0.10.0
pluggy==1.6.0
propcache==0.3.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.13.2
pytest==8.4.1
pytest-django==4.11.1
python-decouple==3.8
python3-openid==3.2.0
pyunormalize==16.0.0
pywin32==311
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rlp==4.1.0
ruff==0.12.5
shelve_cache==0.0.4
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
toolz==1.0.0
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
web3==7.12.1
websockets==15.0.1
yarl==1.20.1
